+++ assets/js/search.js
@@ -1,0 +1,1 @@
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var i=t[r]={i:r,l:!1,exports:{}};return e[r].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var i in e)n.d(r,i,function(t){return e[t]}.bind(null,i));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=309)}({309:function(e,t){summaryInclude=60;var n,r={shouldSort:!0,includeMatches:!0,threshold:.1,tokenize:!0,location:0,distance:100,maxPatternLength:32,minMatchCharLength:1,keys:[{name:"title",weight:.8},{name:"description",weight:.5},{name:"tags",weight:.3},{name:"categories",weight:.3}]},i=(n="s",decodeURIComponent((location.search.split(n+"=")[1]||"").split("&")[0]).replace(/\+/g," "));function a(e){$.each(e,(function(e,t){var n=t.item.description,a="",s=[];r.tokenize?s.push(i):$.each(t.matches,(function(e,t){"tags"==t.key||"categories"==t.key?s.push(t.value):"description"==t.key&&(start=t.indices[0][0]-summaryInclude>0?t.indices[0][0]-summaryInclude:0,end=t.indices[0][1]+summaryInclude<n.length?t.indices[0][1]+summaryInclude:n.length,a+=n.substring(start,end),s.push(t.value.substring(t.indices[0][0],t.indices[0][1]-t.indices[0][0]+1)))})),a.length<1&&n.length>0&&(a+=n.substring(0,2*summaryInclude));var o=function(e,t){var n,r,i,a,s;r=/\$\{\s*isset ([a-zA-Z]*) \s*\}(.*)\$\{\s*end\s*}/g,i=e;for(;null!==(n=r.exec(e));)i=t[n[1]]?i.replace(n[0],n[2]):i.replace(n[0],"");for(a in e=i,t)s=new RegExp("\\$\\{\\s*"+a+"\\s*\\}","g"),e=e.replace(s,t[a]);return e}($("#search-result-template").html(),{key:e,title:t.item.title,link:t.item.permalink,tags:t.item.tags,categories:t.item.categories,description:t.item.description,repo:t.item.repo,registryType:t.item.registryType,language:t.item.language,snippet:a,otVersion:t.item.otVersion});$("#search-results").append(o)}))}i?($("#search-query").val(i),$("#default-body").hide(),function(e){$.getJSON("/registry/index.json",(function(t){var n=new Fuse(t,r).search(e);console.log({matches:n}),n.length>0?a(n):$("#search-results").append("<p>No matches found</p>")}))}(i)):$("#default-body").is(":hidden")&&$("#default-body").show();let s="all",o="all";document.addEventListener("input",(function(e){"componentFilter"===e.target.id&&(o=e.target.value),"languageFilter"===e.target.id&&(s=e.target.value),function(){let e=[...document.getElementsByClassName("component")];"all"===o&&"all"===s?e.forEach(e=>e.classList.remove("is-hidden")):e.forEach(e=>{const t=e.dataset.registrytype,n=e.dataset.registrylanguage;t!==o&&"all"!==o||n!==s&&"all"!==s?e.classList.add("is-hidden"):e.classList.remove("is-hidden")})}()}))}});
