+++ layouts/registry/search.html
@@ -1,4 +1,5 @@
 {{ define "main" }}
+{{ $search := resources.Get "js/search.js" | fingerprint }}
 <div class="registry__search container">
   <div class="row">
     <div class="col">
@@ -28,5 +29,5 @@
 </script>
 <script src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/fuse.js/3.2.0/fuse.min.js"></script>
-<script src="{{ "/js/search.js" | absURL }}"></script>
+<script src="{{ $search.RelPermalink }}" integrity="{{ $search.Data.Integrity }}" type="application/javascript"></script>
 {{ end }}
