+++ src/js/app.js
@@ -1,0 +1,12 @@
+const tracer = require('./tracing');
+
+function navbarToggle() {
+  $('.navbar-burger').click(function() {
+    $('.navbar-burger').toggleClass('is-active');
+    $('.navbar-menu').toggleClass('is-active');
+  });
+}
+
+$(function() {
+  navbarToggle();
+});
